Transaction and Prison initial setup
- Prison and Transaction django models / migrations
- basic transactions endpoint returning list of transactions
- test data
- new load_test_data command that can be used to load test data in dev mode
- unit tests

The model/endpoint support multiple uploads. The endpoint
defines a query param `upload_counter` that can be set to filter by
specific upload number. If not specified, the latest upload will be used.

Note: the endpoint does not perform any authentication and does not
filter by prison at the moment. It just returns all the transactions
belonging to a specific upload batch.