Adds auth to transaction API
This now limits the transactions a user can see
to their own prison. Unless they're a Django admin in
which case they can see everything.

Adds initial groups fixture for each user type but we haven't
mapped exactly what permission each group has, we want to do
this because each user type has a subset of permissions.
Roughly we want NOMSClerk users to be able to mark
transactions as credited and see a list of them for their prison.
We want FinanceUsers to be able to upload a list of transactions
using the BAI2 file they download from online banking.