Remove prisoner name and randomise prison name/id
This removes the prisoner_name and all the code related to it as we will probably not have it. We don't want to store itunless it's really necessary as sometimes it's classified information.

This also randomises the prison name and id in the test fixtures to avoid any confusion/error in case we end up using it in staging/live.