Separate validation tests into different methods
Not really my way of doing things, I like the old way as if something is
broken with the validation, you only see one failure and when you fix that
you usually fix most of the others. Also, it's easy to add extra checks, even
silly ones, which is good IMHO.

With this new version, you see individual failures which other developers
prefer. It also seems clearer.

I don't mind, so this is it...