Add reconciliation tracking to transaction model
This consists of a new field, reconciled, which is orthogonal to
the status of the transaction. It is now possible to filter by
this field when querying the bank_admin endpoint. A new Log record
is created whenever a transaction is reconciled, which is done
by a patch request to the bank_admin transaction endpoint.