Create a File model which tracks transactions included in downloaded files
The idea is that whenever a client application creates a file
containing a certain set of transactions, it will send a request
recording that these transactions have been included in
the given generated file. This is then used to track which
transactions are remaining to be processed with regards to that
file type.