Allow for the creation of GOV.UK Pay transactions
This allows for the creation and processing of transactions
appropriate for integrating the GOV.UK platform. There is a
new category for such payments, and it is possible to mark the
current status of the payment, such that it will not be processed
until the payment has actually been taken.