Use UTC when manually parsing dates, to match TZ used by default
Transactions in the DB are stored with a recevied_at date with a UTC
timezone, and date params parsed by django-reset-framework are similarly
given a UTC param. In places where this parsing is done in our code,
we have been using the current timezone instead, leading to inconcistencies
and incorrect behaviour. This fixes any manual parsing of a pure date into
a datetime to use the UTC timezone.