Create ref codes on reconciliation instead of creation
This is due to the existence of unidentified transactions - those
which cannot be credited or refunded. As whether or not a transaction
can be credited is not known at the time of creation, we need to defer
the creation of ref codes until the point of reconciliation, after which
we know that the validity of the transaction will not change.