Make newly created prisoner locations inactive, and activate later
When new prisoner locations are created they are inactive. A later
call to the delete_old endpoint will then delete all active locations,
and activate all inactive locations. This allows for a client to upload
new prisoner locations in batches and then switch from the old locations
to the new ones atomically, avoiding any possibility of a half-update
occurring.