Add security profiles
This adds a new application which generates and maintains
sender and prisoner profiles for the purpose of allowing security
users to investigate behaviour. This is intended to replace the
current practice of dynamically generating the grouped data on the
relevant credit endpoints.

Instead, the security profile update will run once a day, updating
all relevant sender and prisoner profiles with reference to the new
credits which have been added since it was last run.