Create ProcessingBatch model to track progress of async credit processing
Due to the slower performance of crediting against the NOMIS API, we
need to support asynchronous processing of credits. The new model allows
the client to define a batch that is being processed, and the status of
the included credits can then be checked to find how many have been
processed.