Group sender profiles by postcode as well as card details
It has been found that the last 4 digits and expiry date of a card
are insufficient to uniquely identify a card, so we will also be
grouping based on the normalised postcode of the billing address.
Although this is likely to incorrectly fragment some sender profiles,
this has been decided to be the lesser of two evils.