Create notification system for triggering events based on data patterns
There are a set of rules defined which can be parameterised and
subscribed to by users, which will then generate events when
records are created which match the rule. The list of notification
events is then accessible to the user.

Rules and Inputs are reference data used to tell the client what
kind of subscriptions can be created. The Subscriptions are what
trigger events, with Parameters to configure them.