Prevent the creation of empty payment batches due to race condition
When reconciling payments, a payment batch object is created and
all payments in the batch are assigned to it. We use select_for_update
on the set of payments that will comprise this batch in order to
avoid interleaving operations, however as we did not evaluate the
queryset before creating the payment batch, this was resulting in new
payment batches being created which then didn't have any payments in.

By changing the conditional from a `count()` to a `len()` this ensures
that the update set is evaluated and a lock acquired before the payment
batch record is created, which should avoid superfluous payment batches
being made.