Error only if user admin tries to change prisons
Before: the code was buggy if the user was part of 2 groups,
"Security" and another group != "UserAdmin". The API endpoint
was not allowing this user to change his/her prisons.

After: the code is now checking for specific groups and only
allowing the update if the user is part of "Security" but
not "UserAdmin".