Add 'Exported at' column to each record in AP export
The data team have requested that each row contains a date field that
includes the date that the record was exported at.

We discusssed alternatives such as getting the time from the file meta
data or the filename (which will include a timestamp) but an additional
field was preferred.

The time is stored when the class is initiated so that all the records
have the same data at the time the class was called, rather than an
inline time when the row was processed, because the data team are
already expecting the time to be identical for all records and the
time that a row was added to the export file didn't feel like it had any
value at all in reality.