Refactor file handling in Analytical Platform dump
In working with the `dump_for_ap` and uplaoding the results of the
export command, we ran into problems with the last 8kb of the exported
data remaining in the buffer when the file was worked with in the
command itself.

Although the final implementation will not be called while any export
data remains in the buffer, this was an unexpected complexity that could
be avoided.

Given that a `with` block closes a file when its finished, we think this
simplifies the export itself. Removing the `argsparse.FileType` however
meant that we didn't have ahook to stdout, so the tests had to update
to use temporary files.

This should also make the conversion to a JSON Lines format easier, as
we need to work with the data export without `csv.DictWriter` help.