Add job info endpoint for Intelligence app users
Jira: https://dsdmoj.atlassian.net/browse/MTP-1266

To allow better selection of user research candidates, it will be very
useful to know what a user's job title is, the area of the prison
estate they work in and some details about their day to day tasks.

The job information has been linked to a user and saved in its own
table to keep the user table cleaner and make this easier to remove if
we decide this information isn't as useful or not needed any longer.

The information is not yet editable and only the POST endpoint is
available to keep this code simpler, although editing the information in
the future is planned so the API endpoint will be opened to PATCH
requests in the future.

We considered creating a new app for the Job Information but felt this
wasn't justified as it would be really small. It also felt better to
build on the existing tests around user records than replicate lots of
test set up that already exists.

Co-authored-by: Matt Bee <matt.bee@madetech.com>