Add JobInformation to django admin portal
Jira: https://dsdmoj.atlassian.net/browse/MTP-1266

This commit adds the Job Information to the Django admin interface,
as well as extending the admin view to include user fullname and email.

Given the standard nature of this, we didn't feel tests added any value
and would just be testing that the Django admin 'works'.

Co-authored-by: Matt Bee <matt.bee@madetech.com>