Create convenience methods for uploading export to AP
This commit adds `upload_dump_for_ap` and `dump_and_upload_for_ap`
commands.

When uploading the data dump to the analytical platform (AP) s3 bucket
we assume a role using aws sts which returns a set of temporary security
credentials that you can use to access ap's s3 bucket.

We are using the boto3 library as it makes working with AWS really
simple and is the preferred choice of AWS themselves[1]

The upload won't work locally as the environment variables defined in
base.py need to be set and a working ARN (`AWS_IAM_ROLE_ARN`) role
available, which are only currently set in test and prod environments.

ExtraArgs need setting on the `upload_file` call in
`upload_file_to_analytical_platform` even though the file should
encrypted at rest, it doesn't work unless the explicit arguments are
set.

1. https://aws.amazon.com/sdk-for-python/