MTP-1393 Swagger Implementation for test and dev environments
MTP-1393 Initial commit of swagger implementation. Working in dev environment except for authentication

There are a fair few django-esque normalisations of API implementation that have fallen out of this.
The fact that django-rest-framework won't accept introspection of a APIView/APIViewSet without a serializer is annoying, but I think that `NullSerializer` is probably the nicest possible way of handling this

The way that django-rest-framework handles action mappings in introspection is very strange, I'm not sure if there's a nicer way to do it than changin the APIView to an APIViewSet. To be honest i'm not even sure if it made sense to start with to have an action attribute on an APIView rather than an APIViewSet

MTP-1393 Experimental bash > 4.0 fix for lowercase coercion of branch names in circleci

MTP-1393 Reverted previous bash > 4.0 fix for one compatible with all bash versions

MTP-1393 Reverted premature change of APIView to GenericViewSet, instead subclassed the Django Rest Framework SchemaGenerator to add support for custom actions

MTP-1393 Partial styling fixes to pass CI

MTP-1393 Moved from django-rest-framework-swagger to drf-yasg as the former doesn't work nicely with Oauth2

django-rest-framework-swagger logs in fine with Oauth2, but does not include the returned token in any future requests, or refresh the view to reflect the updated permissions. Also it doesn't use swagger v3
drf-yasg has shown up instances where we break with django best-practises when it comes to serializers. We create different serializers to accomodate differrent RESTful operations (e.g. Update vs Create), rather than including both methods in the same serializer and handling restriction to one code flow through each view via permission_classes. I've worked around thi s for now, but we should fix this at some point

MTP-1393 Ensured that the documentation is only accessible on test and locally, and added clauses necessary for DRFs schema generator to introspect views with request logic

MTP-1393 Changes arising from PR comments