Run Check against created credit in load_test_data
To better test the FIU pending credits page, it would be useful to
have a set of checks to display on the pending credits page.

While this will create a bank of automatically accepted Checks, digging
through the test data to create checks that are placed in the delayed
capture state was proving difficult, so we can come back to that.

To make these Checks show in the credits to action page for FIU we can
update their state in Django admin for the time being.