MTP-1352 Enforce consistency in naming across sender/prisoner profiles, even though it is not stored on profiles
We have to support the edge case of Sender/PrisonerProfile not existing for mtp_api.apps.core.tests.test_tokens.TokenRetrievalTestCase.test_resetting_data_creates_special_user
Additionally, we ensure that the ScheduledCommand tests don't swallow Exception level log statements, otherwise it becomes very difficult to debug their failures as if you have `suppress_logging` wrapping a `ScheduledCommand` invocation in a test case, and that command raises an exception, that exception will be caught by the try-except `Exception` clause within `run_scheduled_commands.py`, and be logged by the `logger.exception` statement. But since this log statement has level `ERROR`, and we suppress everything under `CRITICAL` that log message will be swallowed by `supresss_logging`. "But", you say, "can't you just inspect the database in a post-mortem pdb on a test assertion raising to figure out why the job failed". Unfortunately not, as the exception, though caught, propagates up outside the scope the django `atomic` decorator, causing django to revert the entire transaction (including the test setup), so you'll be left with a blank database, which is why the test assertion raised in the first place