Filter prison list for balances
While we are using the manual fallback for managing prisoner balances
for prisons that don't use nomis for balances. Initially we could only
order the prisons we are uploading data for by private prisons (most
used) to make the UI easier.

In the previous commit we added the `use_nomis_for_balances` column,
so now we can filter to only prisons we should be adding this data for,
reducing risk of choosing the wrong prison and adding an extra layer of
validation.