admin: Removed 'Load Offenders' functionality
This was using the `load_prisoner_locations` command which was retrieving
prisoners locations from Single Offender ID service.

This admin functionality is not used as we don't use the SOID service.

The view and the corresponding form/template are gone.
Probably the most "interesting" thing removed is the
`prison/prisonerlocation/change_list.html` template which was used to add
a 'Load offenders' "pill button" alongside the 'Add prisoner location' one.

Part of ticket: https://dsdmoj.atlassian.net/browse/MTP-1672