Deleted load_prisoner_locations command
This command (in `prison` app) was using Single Offender ID service to
retrieve prisoners locations. This service is no more so this can go.

As a result the `OFFENDER_API_*` settings can also go as these contained
the SOID service endpoint and credentials and they're not used anywhere
else now (in fact they may not even be set in production).

Part of ticket: https://dsdmoj.atlassian.net/browse/MTP-1672