Removing SOID migrations (so that these columns are not deleted)
The `single_offender_id` column was removed from 3 tables/models.

This may cause problems in production so to we're thinking a good
approach to this would be to delete these migrations (so that the
columns are not dropped in production) and once this is released to
production a separate PR with these migrations added back can be
released when no code is using these fields.

Slack conversation: https://mojdt.slack.com/archives/C015WEVRQSX/p1612439976010100?thread_ts=1612438739.002800&cid=C015WEVRQSX