CI: Run Django checks and check if code misses migrations
Running some Django checks as part of our CI build pipeline:
- `./manage.py checks` and
- `./manage.py makemigrations --check` which would return a non-zero status
  when new migrations were generated. e.g. potentially we forgot to run
  makemigrations and we didn't include them

This additional `run_checks` job will run alongside the `test` job
as it should be relatively fast and it's useful to know as soon as possible
if we missed something in a PR for example.

**NOTE**: Also note that this would not block push of the docker image or
deploy to test.
The assumption is that sometimes it may make sense to temporarily leave
some DB schema changes/migrations out of a PR.

If we later on decide this checks can be blocking we can always change
the CI workflow to be more draconian.

Ticket: https://dsdmoj.atlassian.net/browse/MTP-1796