load_test_data/dev-prison-api added more prisons
We found another 2 prisons present in the dev version of HMPPS Prison API
so we tweaked the logic to generate PrisonerLocations from 3 prisons.

The code tries to generate a balanced number of prisoner locations from
these prisons as much as possible with the returned data (which currently
is high anyway, but this may change in the future).