MTP-1813: prison optional in AccountRequest model
In `noms-ops`' sign-up form we don't ask for the 'Prison' so this needs
to be blank in DB so that 'security' AccountRequest can be persisted even
if they have an empty 'prison' value.

Validation will be added to check it's still present for non-'security'
account requests.

Part of ticket: https://dsdmoj.atlassian.net/browse/MTP-1813