send_account_request_emails works with requests without prison
'security' account requests are not prison-specific.
These 'security' account requests are emails to all the 'security' UserAdmins
(FIU), regardless of prison associated to these UserAdmins.

Test for this command have been improved to separate the test for the
`find_admins()` method from the test which checks correct admins are emailed.

NOTE: The `mtp_auth/new_account_requests.(html|txt)` templates don't seem
to use the `prison_name` in the context so I've removed it.