Make 'security' user a UserAdmin makes them FIU
In the context of intelligence tool, FIU users are equivalent to user admins.
If a security user is in `FIU` group they would also be in `UserAdmin`
group and vice versa.

The helper method `_make_user_admin()` was called just after the user's
groups were clear which meant the user was not added to `FIU` group as
intended.

See comment with more links to code: https://github.com/ministryofjustice/money-to-prisoners-api/pull/647#discussion_r615735071