Made check's rules tests more robust
Some of these tests have been flaky.
This is caused by the random nature of the payments generated in the
test setup.

Sometimes these payments may trigger frequency rules in addition to
other rules expected to be triggered.
In order to avoid this, these tests have been tweaked to assert for inclusion
(`assertIn()`) rather than testing for lists equality.

This makes these tests less strict but the intent of the test should
be maintained.

Part of: https://dsdmoj.atlassian.net/browse/MTP-1370