More fixes for PrivateEstateBatch flaky test
Problem
-------

One of the tests was still calling `date()` on a datetime coming from the
database (UTC) and then comparing with the batch dates which are calculated
on localtime dates now.

This was sometimes causing a date mismatch when some of the credits
were received near midnight as their date would erroniously be the day
before, e.g.

1. Credit `received_at` 23:30 UTC of 5th May 2021
2. This is 00:30 BST (+01:00) of the 6th May 2021
3. This credit belongs to batch 6th May 2021
4. but just use `.date()` on UTC `received_at` would return 5th May 2021
   causing the test to fail

Circle CI example failure: https://app.circleci.com/pipelines/github/ministryofjustice/money-to-prisoners-api/946/workflows/d7f5100c-30cd-40f5-bd57-21256d16a930/jobs/1986

Solution
--------
Get the date off the `received_at` datetime of a credit using
`django.utils.timezone.localdate()` this would return the local date
honouring the timezone.

Assumption
----------
Batches for a date are for the credits in the day from midnight to midnight
of the local time (Europe/London) rather than midnight to midnight UTC.

Other changes
-------------
In the tests I've also moved the calculation of `date_with_batch` into
the `setUp()` method as this value was used in almost all the tests anyway.

This is also using the local date rather than the date from the UTC time.