Added index to transaction table
Transactions are not used as much as before but noms-ops Credits search
still allows users to search by these fields and they could be slow
without these indexes.

The most important of these indexes is probably the one on `sender_name`
which is currently searched when a user use the `noms-ops` simple search.

Tickets:
- generic: https://dsdmoj.atlassian.net/browse/MTP-1523
- search: https://dsdmoj.atlassian.net/browse/MTP-1273