Add methods to user satisfaction models and queryset to calculate the percentage of responses that were satisfied or very satisfied This mirrors the methods of digital take-up model and queryset.
NB: This it NOT how GOV.UK Performance Platform used to calculate a "satisfaction" percentage. They arbitrarily assigned 0% 25% 50% 75% and 100% to the 5 ratings and performed an average of the results. Statistically speaking, this is incorrect as you cannot meaningfully turn categorical data into a continuous number. For example: if all responses were "Dissatisfied", their calculation would result in 25% satisfaction, however does that not imply that a quarter of all people were satisfied? In fact nobody was satisfied and yet nobody was _very_ dissatisfied.

We use a statistically-meaningful calculation: simply the percentage of responses that were _either_ satisfied _or_ very satisfied. It does not show how very satisfied or dissatisfied people were, but that cannot be done with a single figure.