Added headers to /performance/data endpoint
This dictionary contains the Performance Data fields and their
verbose name to be used by client to construct the headers row of the
resulting CSV file.

This means that potential addition to the PerformanceData model would be
included in the CSV file without having to update the `send-money` service.

Part of: https://dsdmoj.atlassian.net/browse/MTP-1893