update_performance_data command defaults to 2 weeks ago
When run without passing --week-from and --week-to the command defaults
to calculate the perfomance data values for 2 weeks ago.

The assumption is that data needed to populate PerformancePlatform
records may be delayed but we'd expect this delay to be less than 2 weeks.

This is likely the way the scheduled command will be run.

Part of ticket: https://dsdmoj.atlassian.net/browse/MTP-1895