Update Performance data on new User Satisfaction data
Some of the Performance data is related to the User Satisfaction data
which may not be available on a weekly basis.

Performance data is updated weekly even when part of the data may not
yet be available.

With this commit when new User Satisfaction data is uploaded the
`update_performance_data` command is run (asyncronously) for the relevant
weeks.

NOTE: We're assuming the uploaded data is for continuous days.
When the last uploaded date is a Sunday we assume this is a complete week
(or it's completing a previously uploaded week) so the `update_performance_data`
command is run to update this last week.

Part of ticket: https://dsdmoj.atlassian.net/browse/MTP-1899