Add management command to mark "stuck" credits as credited When the cashbook posts transactions to NOMIS, it occasionally fails to update a batch as credited in the MTP api. This means that credits correctly reach NOMIS (and the prisoner) but the MTP api is in an inconsistent state and the cashbook continues to display them as pending new credits.
The NOMIS api (theoretically) will not allow double-posting a transaction if someone retries to credit these in the cashbook.

This command matches old credits in the MTP DB that are thought to still be pending with transactions in NOMIS: if there is a one-to-one mapping and the NOMIS transaction IDs hadn't been used before, this command offers to link them and fix the problem.