Fix date parsing in tests (`DATETIME_INPUT_FORMATS` change)
Parsing of some of the dates was failing because the `DATETIME_INPUT_FORMATS`
setting no longer includes plain dates formats (e.g. `'%Y-%m-%d'`).

- `DATETIME_INPUT_FORMATS` in Django 2.2:
  - https://docs.djangoproject.com/en/2.2/ref/settings/#std:setting-DATETIME_FORMAT
- `DATETIME_INPUT_FORMATS` in Django 3.2:
  - https://docs.djangoproject.com/en/3.2/ref/settings/#std:setting-DATETIME_FORMAT

I've added `DATE_INPUT_FORMATS` to the list of formats attempted in the
test.

Documentation for `DATETIME_INPUT_FORMATS` in Django 3.2 says:

> Date-only formats are not included as datetime fields will automatically
> try DATE_INPUT_FORMATS in last resort.

So maybe this is just a byproduct of the specific test and the use of this
setting directly to try and parse strings as dates?

Interestingly this change is not mentioned in the Django 3.2 release
notes.