Fix for Django admin tick chart style
Some of the admin site pages show a chart above the results (e.g.
to show the trend of credits), after the move to Django 3.x this chart
was showing alongside the results rather than on top breaking the style
of the page (as would make the table narrow)

I've changed the block where the chart sits and it's now showing in the
same position as it was: I've moved it on top of the `result_list` block
which makes sense.