MTP-1970: Added (concurrent) index on Credit's `created` column
We can't rely on `received_at` to determine if a Credit is older than
7 years. This is because a Credit may never be received, and this column
may be `NULL`.

The `created` column could be used for this purpose, but in order to do
so in a timely manner we need to add an index on this column.

The index is added using Postgres's `AddIndexConcurrently` which adds
the index without blocking write operations on the `credit_credit` table.

The Postgres documentation explains this in details, but effectively
the index is created in an `INVALID` state while the table is scanned twice
and then eventually marked as `VALID` (when RDMS starts to use it).

**NOTE**: Potentially the creation of the index could fail, in that case
the index would sit there in `INVALID` state and not used when querying
(but would still be updated when somethin changes). It should hopefully
be less likely to fail as this is not a unique index.

See:
- [`AddIndexConcurrently` Django docs](https://docs.djangoproject.com/en/3.2/ref/contrib/postgres/operations/#concurrent-index-operations)
- [Postgres documentation on adding indexes concurrently](https://www.postgresql.org/docs/current/sql-createindex.html#SQL-CREATEINDEX-CONCURRENTLY)