MTP-1970: Remove old SavedSearch
The relationship between SavedSearch records and the corresponding
`PrisonerProfile`/`SenderProfile` is an implicit one that the RDMS isn't
aware of.

These records would not be CASCADE-deleted automatically when
the corresponging `PrisonerProfile`/`SenderProfile` is deleted.

This not only leaves orphan/junk data in DB table (these records
contains some personal information from public users and should
be deleted) but most importantly they could cause exceptions
when a security user would try to display a page with these searches
(because of the broken references to now-deleted data)

The `delete_old_data` command now explicitly search for SavedSearch
records for the corresponging `SenderProfile` or `PrisonerProfile` by
looking into the `site_url` fields (this search doesn't uses any
available indexes but this table contains a very small amount of
records so that is not problematic)