MTP-1970: Show deletion/updates tree when logging
Rather than show all log lines at the same level in a flat manner,
The `delete_old_data` command now shows these logs in a nested way.

This is useful to see the relationship between the deleted/updated
records and what triggerred them.

For example:

```
Deleting Credits older than 2016-01-03...
	[...]
	Deleting Credit Credit 4, £40 Mrs. Walls > JOHN WALLS, credited...
		Deleting associated Events...
			Deleting Event Event object (5)...
				Records deleted: (2, {'notification.CreditEvent': 1, 'notification.Event': 1}).
		Records deleted: (3, {'payment.Payment': 1, 'credit.Log': 1, 'credit.Credit': 1}).
		SenderProfile Sender 2 has no credits, deleting...
			Deleting associated Events...
				Deleting Event Event object (10)...
					Records deleted: (2, {'notification.SenderProfileEvent': 1, 'notification.Event': 1}).
			Deleting SenderProfile's SavedSearches '*/senders/2/*'...
				Records deleted: (1, {'security.SavedSearch': 1}).
			Records deleted: (3, {'security.SenderEmail': 1, 'security.DebitCardSenderDetails': 1, 'security.SenderProfile': 1}).
		PrisonerProfile Prisoner 2 (B1510BF) updated.
    [...]
```

By looking at this output it's clear - even without knowing much about
MTP data model - that Credit 4 was deleted and that triggered the deletion
of the related events/payments.
The SenderProfile associate with that credit was also deleted because
it has no other credits. This in turn caused the deletion of events,
saved searches and other related records. But also cause triggered the
update of the PrisonerProfile 2 (its counts for credits could have changed)

Hopefully this will make it easier to follow the chain of deletions both
when we'll test this against production data but also when this will run
in production.