MTP-1970: Check for orphan bank accounts when deleting sender profiles
When deleting senders profiles (because they don't have any credits),
any potential linked `BankTransferSenderDetails` would also be
CASCADE-deleted.

These `BankTransferSenderDetails` records were from the time when
the public could still send money to prisoners using bank transfers.

However deleting a `BankTransferSenderDetails` could lead to an
ophan `BankAccount` because these are not CASCADE-deleted - and they
should not because a bank account could be linked to multiple
senders or recipients beside the one being deleted.

When deleting a `SenderProfile` the `delete_old_data` now checks if
the linked bank accounts still have any recipients or senders, and if not
delete them as they're orphans.