MTP-1970: Logic to delete orphan Billing Addresses
A `BillingAddress` is linked to a `SenderProfile` via the
`DebitCardSenderDetails` model.

When a `SenderProfile`/`DebitCardSenderDetails` is deleted this link is
broken (`BillingAddress`'` `debit_card_sender_details_id` is set to `NULL`).

(Card) `Payment`s also have a `BillingAddress`.

When an old `Credit` is deleted the `Payment` is also deleted and
`BillingAddress` may become orphan (no link with `DebitCardSenderDetails`
nor `Payments`).

Before deleting the credit the `delete_old_data` command take note of the
`BillingAddress` (if there is one) and then after the Credit/Payment is
deleted it checks if this billing address has `debit_card_sender_details_id`
set to `NULL` and no linked `Payment`s, if this is the case it means
it should be safely deleted.